requirements.txt
python-dotenv==0.19.1
requests
selenium
python-dotenv
python-decouple
bs4